// src/types/rayso.ts
export type RaysoTheme = {
  light: RaysoBlock;
  dark: RaysoBlock;
};

export type RaysoBlock = {
  text: string;
  text_2: string;
  text_3: string;
  interface: string;
  interface_2: string;
  interface_3: string;
  background: string;
  background_2: string;
  primary: string;
  secondary: string;
  accent: string;
  accent_2: string;
  accent_3: string;
};